Clang-Tidy (#3176)
Add a configuration file .clang-tidy for clang-tidy. Add support to it
in CMake. It can be enabled by `-DAMReX_CLANG_TIDY=ON`.

Fix many warnings raised by clang-tidy. Most of them are fixed by
`clang-tidy --fix-errors`.

To bypass clang-tidy, some unfixable codes have been moved to
`*.nolint.H`.

A number of issues have been found and fixed.
    
* There was use-after-move in testing OpenMP thread safety for
communication.

* The MultiFab tags were not handled correctly during vector resize.

* Parentheses were missing in the definition of AMREX_REAL_LOWEST.

+++ Src/Amr/AMReX_AuxBoundaryData.cpp
@@ -63,7 +63,7 @@ AuxBoundaryData::initialize (const BoxArray& ba,
 
     const bool verbose   = false;
     const int  NProcs    = ParallelDescriptor::NProcs();
-    const auto strt_time = amrex::second();
+    const auto strt_time = amrex::second(); // NOLINT
 
     m_ngrow = n_grow;
 
@@ -109,7 +109,7 @@ AuxBoundaryData::initialize (const BoxArray& ba,
     {
         const int IOProc   = ParallelDescriptor::IOProcessorNumber();
         auto      run_time = amrex::second() - strt_time;
-        const int sz       = nba.size();
+        const int sz       = static_cast<int>(nba.size());
 
 #ifdef BL_LAZY
         Lazy::QueueReduction( [=] () mutable {
